Update README.md
To understand the genuine capability of computerized promoting, advertisers should dive profound into the present immense and complex deals channel to find techniques that can have an effect. Digital marketing is a way to effectively interact with existing customers and gather real-time data on potential customers. Businesses increase brand recognition, establish themselves as thought leaders in the industry, and position themselves at the forefront of the Internet when customers are ready to buy by engaging customers in the digital environment. A professional with multiple skills who uses all digital marketing channels to promote a business or product online is called a digital marketing specialist. They must be well-versed in all areas and tools of digital marketing. Job Responsibilities in